requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
email-validator==1.1.3
Flask-Migrate==4.0.5
pymysql==1.1.0
python-dateutil==2.8.2
bcrypt==4.0.1
